+++ src/Mobs/Monster.h
@@ -277,13 +277,13 @@ protected:
 	void AddRandomUncommonDropItem(cItems & a_Drops, float a_Chance, short a_Item, short a_ItemHealth = 0);
 
 	/** Adds one rare item out of the list of rare items a_Items modified by the looting level a_LootingLevel(I-III or custom) to the itemdrop a_Drops*/
-	void AddRandomRareDropItem(cItems & a_Drops, cItems & a_Items, short a_LootingLevel);
+	void AddRandomRareDropItem(cItems & a_Drops, cItems & a_Items, unsigned int a_LootingLevel);
 
 	/** Adds armor that is equipped with the chance saved in m_DropChance[...] (this will be greter than 1 if piccked up or 0.085 + (0.01 per LootingLevel) if born with) to the drop*/
-	void AddRandomArmorDropItem(cItems & a_Drops, short a_LootingLevel);
+	void AddRandomArmorDropItem(cItems & a_Drops, unsigned int a_LootingLevel);
 
 	/** Adds weapon that is equipped with the chance saved in m_DropChance[...] (this will be greter than 1 if piccked up or 0.085 + (0.01 per LootingLevel) if born with) to the drop*/
-	void AddRandomWeaponDropItem(cItems & a_Drops, short a_LootingLevel);
+	void AddRandomWeaponDropItem(cItems & a_Drops, unsigned int a_LootingLevel);
 
 
 } ;  // tolua_export
